+++ Shabok/src/main/java/com/shabo/api/controllers/LibroController.java
@@ -38,7 +38,7 @@ public class LibroController {
 	}
 	
 	@GetMapping("/libros/{id}")
-	public Libro mostrarLibroID(@PathVariable(name="id") int id){
+	public Libro mostrarLibroID(@PathVariable(name="id") long id){
 		return librosServiceImpl.mostrarLibroID(id);
 	}
 	
@@ -48,7 +48,7 @@ public class LibroController {
 	}
 	
 	@PostMapping("/libros/modificar/{id}")
-	public Libro modificarLibro(@PathVariable(name="id") int id, @RequestBody Libro libro) {
+	public Libro modificarLibro(@PathVariable(name="id") long id, @RequestBody Libro libro) {
 		Libro libroSeleccionado = new Libro();
 		Libro libroModificado = new Libro();
 		
@@ -61,7 +61,7 @@ public class LibroController {
 	}
 	
 	@DeleteMapping("libro/eliminar/{id}")
-	public void eliminarEquioi(@PathVariable(name="id") int id) {
+	public void eliminarEquioi(@PathVariable(name="id") long id) {
 		librosServiceImpl.eliminarLibro(id);
 		
 	}
