+++ Shabok/src/main/java/com/shabo/api/dao/ICategoriaDAO.java
@@ -11,6 +11,6 @@ import com.shabo.api.dto.Categoria;
  * @author Fenrir
  *
  */
-public interface ICategoriaDAO extends JpaRepository<Categoria, Integer> {
+public interface ICategoriaDAO extends JpaRepository<Categoria,  Long > {
 
 }
